pyproject.toml
[tool.poetry]
name = "vinted-bot"
version = "0.1.0"
description = ""
authors = ["Dincio"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^21.1.1"
selenium = "^4.20.0"
webdriver-manager = "^4.0.1"
fire = "^0.6.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
pyimagine = "^3.3.8"
wget = "^3.2"
transformers = "^4.41.0"
torch = "^2.3.0"
pillow = "^10.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
